Move CLI version exceptions to dandi/exceptions.py
