+++ dandi/exceptions.py
@@ -23,3 +23,34 @@ class FailedToConnectError(RuntimeError):
     """Failed to connect to online resource"""
 
     pass
+
+
+class CliVersionError(RuntimeError):
+    """ Base class for `CliVersionTooLowError` and `BadCliVersionError` """
+
+    def __init__(self, our_version, minversion, bad_versions):
+        self.our_version = our_version
+        self.minversion = minversion
+        self.bad_versions = bad_versions
+
+    def server_requirements(self):
+        s = f"Server requires at least version {self.minversion}"
+        if self.bad_versions:
+            s += f" (but not {', '.join(map(str, self.bad_versions))})"
+        return s
+
+
+class CliVersionTooLowError(CliVersionError):
+    def __str__(self):
+        return (
+            f"Client version {self.our_version} is too low!  "
+            + self.server_requirements()
+        )
+
+
+class BadCliVersionError(CliVersionError):
+    def __str__(self):
+        return (
+            f"Client version {self.our_version} is rejected by server!  "
+            + self.server_requirements()
+        )
